requirements.txt
untangle
hickle
python-docx
wavedrom
wavedrom-ascii
# parsers
cmsis-svd
ipyxact
peakrdl-ipxact
peakrdl-verilog
peakrdl-uvm
peakrdl-cheader
# transport layers
# pylink-square
# pyocd # requires cargo and rust toolchain
# esptool
# pyverilator-mm
# other
# packaging # for verilator version check
